Add package which exposes nodeinfo as JSON over HTTP
When a node is registering with the Freifunk Stuttgart network, the
onboarder requests nodeinfo, such as the ZIP code, from the node to be
registered. It does so by sending an HTTP request to /cgi-bin/nodeinfo,
where it expects to find nodeinfo as JSON.

In Gluon v2018.1, the status page was rewritten [1] and the nodeinfo API
has been removed.

This commits adds the nodeinfo API as in previous releases. To adress
privacy issues of the old status page, CORS headers are not send. The
API is therefore not accessible by foreign web pages.

[1] <https://gluon.readthedocs.io/en/v2018.1.x/releases/v2018.1.html#status-page-rewrite>